Added default files, fixed tiny warp list glitch, and updated plugin.yml
This is likely going to be the last "major" update I push to Fusion
until it either goes public on GitHub, sold off to someone or abandoned
like most of my projects. Although being dramatic isn't my thing and no
one is going to read this, this project has been the biggest and most
important project I've worked on and I'm glad I've seen it through to
something that I could've have imagined at the start. My inability to
finish projects is a major issue with my development work and
personality in general, but this is the one thing that I have completed
purely for the love of programming. If this project goes open-source, I
hope that at least one person attempts to study the code-base, pointing
out its strengths and weaknesses in the design and structure of the
project. This has been a learning experience for me and hopefully to the
others who have dedicated time to work on this plugin. Overall, this
plugin represents the combined experience of at least four people over
the years of playing Minecraft, and its lack of perfection shows just
how much their is to learn in the future. But it's also a reminder of
how far we have all come as developers.

- Guggle